+++ src/Inline/Parser/EntityParser.php
@@ -36,7 +36,7 @@ class EntityParser extends AbstractInlineParser
      */
     public function parse(InlineParserContext $inlineContext)
     {
-        if ($m = $inlineContext->getCursor()->match('/^' . RegexHelper::REGEX_ENTITY . '/i')) {
+        if ($m = $inlineContext->getCursor()->match('/^' . RegexHelper::PARTIAL_ENTITY . '/i')) {
             $inlineContext->getContainer()->appendChild(new Text(Html5Entities::decodeEntity($m)));
 
             return true;