+++ src/Util/LinkParserHelper.php
@@ -27,10 +27,10 @@ final class LinkParserHelper
      */
     public static function parseLinkDestination(Cursor $cursor)
     {
-        if ($res = $cursor->match(RegexHelper::getInstance()->getLinkDestinationBracesRegex())) {
+        if ($res = $cursor->match(RegexHelper::REGEX_LINK_DESTINATION_BRACES)) {
             // Chop off surrounding <..>:
             return UrlEncoder::unescapeAndEncode(
-                RegexHelper::unescape(substr($res, 1, strlen($res) - 2))
+                RegexHelper::unescape(substr($res, 1, -1))
             );
         }
 
@@ -75,8 +75,7 @@ final class LinkParserHelper
      */
     public static function parseLinkLabel(Cursor $cursor)
     {
-        $escapedChar = RegexHelper::getInstance()->getPartialRegex(RegexHelper::ESCAPED_CHAR);
-        $match = $cursor->match('/^\[(?:[^\\\\\[\]]|' . $escapedChar . '|\\\\)*\]/');
+        $match = $cursor->match('/^\[(?:[^\\\\\[\]]|' . RegexHelper::PARTIAL_ESCAPED_CHAR . '|\\\\)*\]/');
         $length = mb_strlen($match, 'utf-8');
 
         if ($match === null || $length > 1001 || preg_match('/[^\\\\]\\\\\]$/', $match)) {
@@ -95,9 +94,9 @@ final class LinkParserHelper
      */
     public static function parseLinkTitle(Cursor $cursor)
     {
-        if ($title = $cursor->match(RegexHelper::getInstance()->getLinkTitleRegex())) {
+        if ($title = $cursor->match('/' . RegexHelper::PARTIAL_LINK_TITLE . '/')) {
             // Chop off quotes from title and unescape
-            return RegexHelper::unescape(substr($title, 1, strlen($title) - 2));
+            return RegexHelper::unescape(substr($title, 1, -1));
         }
     }
 }